Run multiple instances at the same time
This patch will allow MiniDLNA to run multiple instances at the same
time, serving content from different directories.

To use, you need to change:
- Port number
- Friendly name
- UUID

These (except for last) can be done in a separate config file and set
with the -f option.

To make this work, I made the UUID of the server configurable with the
command line option -u. It will only set as much of the UUID as you
provide, for example `-u 5` will use the same UUID, but starting with a
five (instead of four, as the default does).

(Anonymous author: https://sourceforge.net/p/minidlna/patches/11/)